Create README_circle_seq.md
I've been struggling with the original firmware; I never managed to get the recorded CV in tune when it was output. I therefore wanted to try and make the sequencer work somehow, without the need to input cv to make it playable. I first think it should do bass lines, but later I plan on implementing arpeggios also. May be sample and hold or LFO